Make spring-security-oauth2 to use com.friendr:spring-security-jwt
While the project includes local copy of spring-security-jwt the
external version of org.springframework.security:spring-security-jwt
was used as a dependency.

This caused problems with releasing and also restricts possible changes
in codebase.

Now all mentions of org.springframework.security:spring-security-jwt
were removed in favor of internal version of spring-security-jwt.