Add support for JWK Set to be used to sign/verify JWTs
This allows to rotate the keys. The last active key in the set will be
used to sign new minted JWTs. Any valid key in the set will be used to
verify previously minted JWTs.

`JwkSetJwtAccessTokenConverter` should be used instead of
`JwtAccessTokenConverter` to make it work.

The list of changes:
* Add new extension of JwtAccessTokenConverter that uses JWK Set
  instead of a single keypair for JWT signing/verifying.
* Make JwksEndpoint bean to always be created. If no JWK Set is found
  the endpoint will respond with an empty JWKSet.
* JwksEndpoint will filter out JWKs that doesn't match the condition:
  `jwk.nbf <= currentTimestamp < jwk.exp`.
* Add `iat` to claims set of generated JWT.
* Add support of `java.util.Date`, `java.time.Instant` and
  `java.lang.Number` for `iat` and `exp` claims in the claims `Map`.